extract DisplayWatcher to its own File, because it is independent of the WindowManager
