+++ modules/electron/src/WindowManager.ts
@@ -1,12 +1,10 @@
-import type {Display} from '@common/definitions.js';
 import type {BrowserWindowConstructorOptions} from 'electron';
 import type {FrontendIpcListener} from './ipc.js';
-import {EventEmitter} from 'node:events';
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import process from 'node:process';
 import {fileURLToPath} from 'node:url';
-import {app, BrowserWindow, screen} from 'electron';
+import {app, BrowserWindow} from 'electron';
 import {v4 as uuidv4} from 'uuid';
 import {registerFrontendListener, unregisterFrontendListener} from './ipc.js';
 import {getLogger} from './logging.js';
@@ -85,60 +83,3 @@ export async function createWindow(conf: WindowConfiguration): Promise<BrowserWi
     log.error('Error in Function createWindow:', conf, '/n', e);
   }
 }
-
-class DisplayWatcher extends EventEmitter {
-  private static instance: DisplayWatcher;
-  private currentDisplays: Display[] = [];
-
-  private constructor() {
-    super();
-    this.init().catch(reason => log.error('error during DisplayWatcher init', reason));
-  }
-
-  public getDisplays(): Display[] {
-    return this.currentDisplays;
-  }
-
-  public static getInstance(): DisplayWatcher {
-    if (!DisplayWatcher.instance) {
-      DisplayWatcher.instance = new DisplayWatcher();
-    }
-    return DisplayWatcher.instance;
-  }
-
-  private async init() {
-    await app.whenReady();
-    this.startWatching();
-  }
-
-  private startWatching() {
-    const updateDisplays = () => {
-      const primaryDisplayId = screen.getPrimaryDisplay().id;
-
-      this.currentDisplays = screen.getAllDisplays().map(display => ({
-        ...display,
-        primary: display.id === primaryDisplayId,
-      }));
-
-      log.debug('Updated display layout:', this.currentDisplays);
-
-      this.emit('update', this.currentDisplays);
-    };
-
-    updateDisplays();
-
-    screen.on('display-added', updateDisplays);
-    screen.on('display-removed', updateDisplays);
-    screen.on('display-metrics-changed', updateDisplays);
-  }
-
-  public on(event: 'update', listener: (displays: Display[]) => void): this {
-    return super.on(event, listener);
-  }
-
-  public off(event: 'update', listener: (displays: Display[]) => void): this {
-    return super.off(event, listener);
-  }
-}
-
-export const DISPLAY_WATCHER = DisplayWatcher.getInstance();
